Introduce the concept of a "none" value for `OptionalArg`
Sometimes there is a requirement for a parameter that is used
extensively with a regular value, and rather than specifying this on
every single call, we can set this value to be the default and only
explicitly specify it when it needs to be changed. Unfortunately, doing
this makes it tricky to provide a null or empty value for this argument.

Introducing the concept of the "none" value would allow the default to
be used, whilst somewhat conveniently (and explicitly) indicating that a
parameter is being intentionally omitted.